create PortfolioTransaction class add pnl to positions-new api introduce markPrice

+++ src/main/java/tw/util/S.java
@@ -484,7 +484,7 @@ public class S {
 		return sb.toString();
 	}
 
-	public static int between(int num, int min, int max) {
+	public static double between(double num, double min, double max) {
 		return Math.min( Math.max( num, min), max);
 	}
